refactoring: split up traversal types
